Take the buffer lengths in the C API as size_t
The four print-to-buffer entry points returned their length through an
"unsigned long*". That is 64 bits on Linux and macOS but 32 on i386 and
on 64-bit Windows, so the one installed header described a different
function depending on the target.

size_t is what these actually are: every implementation stores
s.size() + 1 into them. It is also the least disruptive of the
consistent choices -- identical to the current type on Linux and macOS
at both word sizes, and on i386. Only 64-bit Windows callers see a
width change, from 32 to 64 bits.

This is an ABI change to installed API, so it needs a version note.

The ctypes bindings declared the same parameter as c_ulong, which had
the matching bug; they now use c_size_t. The malloc failure messages
move from %lu to %zu.

+++ tests/api/C/fp-multi-checker.cpp
@@ -227,7 +227,7 @@ TEST(fp_multi_checker, print_counterexample_across_checkers)
 
   // Print v1's counterexample into a buffer; it must not touch v2's manager.
   char* buf = NULL;
-  unsigned long len = 0;
+  size_t len = 0;
   vc_printCounterExampleToBuffer(v1, &buf, &len);
   EXPECT_TRUE(buf != NULL);
   EXPECT_TRUE(len > 0);
